fix /api route forwarding
