+++ src/app.js
@@ -13,6 +13,7 @@ const { ObjectId } = require('mongoose').Types;
 
 const { BODYPARSER_JSON_LIMIT, LEAD_TIME } = require('../config/globals');
 
+const apiPath = require('./middleware/apiPath');
 const middleware = require('./middleware');
 const setupConfiguration = require('./configuration');
 const sockets = require('./sockets');
@@ -94,6 +95,7 @@ const setupApp = async () => {
 		});
 	}
 	app
+		.configure(apiPath)
 		.configure(services)
 		.configure(components)
 		.configure(sockets)
